Fixed an app breaking typo and moved interfaces back into interfaces package
