+++ src/main/java/com/tapsintown/interfaces/Locations.java
@@ -1,4 +1,4 @@
-package com.tapsintown.models;
+package com.tapsintown.interfaces;
 
 import com.tapsintown.models.Location;
 import org.springframework.data.repository.CrudRepository;
