--- src/main/java/com/tapsintown/models/Users.java
@@ -1,9 +1,0 @@
-package com.tapsintown.models;
-
-import org.springframework.data.repository.CrudRepository;
-
-/**
- * Created by anthonyfortney on 1/17/17.
- */
-public interface Users extends CrudRepository<Users, Long> {
-}
